% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_accounts.R
\name{api_accounts}
\alias{api_accounts}
\title{RobinHood API User}
\usage{
api_accounts(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Returns a list like object of class RobinHood
}
