% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_accounts_crypto.R
\name{api_accounts_crypto}
\alias{api_accounts_crypto}
\title{RobinHood API: User Crypto}
\usage{
api_accounts_crypto(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function which calls the Nummus API to return the account id.
}
