% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ach.R
\name{api_ach}
\alias{api_ach}
\title{RobinHood API: ACH}
\usage{
api_ach(
  RH,
  action,
  amount = NULL,
  status_url = NULL,
  cancel_url = NULL,
  transfer_url = NULL
)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "transfers", "relationships", "schedules", "status", "cancel", "deposit", "withdraw"}

\item{amount}{(numeric) amount in dollars you want to deposit or withdraw (NULL if not one of those actions)}

\item{status_url}{(string) URL returned by place_ach_transfer()}

\item{cancel_url}{(string) URL returned by place_ach_transfer()}

\item{transfer_url}{(string) url of your linked account, output of get_ach(RH, "relationships")}
}
\description{
Backend function for interacting and getting data on linked bank accounts.
}
