% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_contracts.R
\name{api_contracts}
\alias{api_contracts}
\title{RobinHood API: Option Contract Instruments}
\usage{
api_contracts(RH, chain_symbol, type)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{chain_symbol}{(string) a single ticker symbol}

\item{type}{(string) one of call or put}
}
\description{
RobinHood API: Option Contract Instruments
}
