% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_currency_pairs.R
\name{api_currency_pairs}
\alias{api_currency_pairs}
\title{RobinHood API: Currency Pairs}
\usage{
api_currency_pairs(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Returns the ID for a particular paired currency
}
