% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_historicals.R
\name{api_historicals}
\alias{api_historicals}
\title{RobinHood API: Historicals}
\usage{
api_historicals(RH, historicals_url, body)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{historicals_url}{(string) api url}

\item{body}{(string) api body}
}
\description{
Backend function called by get_historicals(). Returns a data frame of historical price data.
}
