% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_historicals_options.R
\name{api_historicals_options}
\alias{api_historicals_options}
\title{RobinHood API: Historicals Options}
\usage{
api_historicals_options(
  RH,
  chain_symbol,
  type,
  strike_price,
  expiration_date,
  interval = NULL,
  span = NULL
)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{chain_symbol}{(string) stock symbol}

\item{type}{(string) one of ("put", "call")}

\item{strike_price}{(numeric) strike price}

\item{expiration_date}{(string) expiration date (YYYY-MM-DD)}

\item{interval}{(string) one of ("5minute", "10minute", "hour", "day", "week")}

\item{span}{(string) one of ("day", "week", "month")}
}
\description{
Backend function called by get_historicals_options().
}
