% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_instruments.R
\name{api_instruments}
\alias{api_instruments}
\title{RobinHood API: Instruments}
\usage{
api_instruments(RH, instrument_id)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{instrument_id}{(string) a single instrument_id}
}
\description{
Backend function called by get_tag(), get_position(), watchlist(). Returns a list of instrument data.
}
\examples{
# data returned by api call
#  $ margin_initial_ratio
#  $ rhs_tradability
#  $ id
#  $ market
#  $ simple_name
#  $ min_tick_size
#  $ maintenance_ratio
#  $ tradability
#  $ state
#  $ type
#  $ tradeable
#  $ fundamentals
#  $ quote
#  $ symbol
#  $ day_trade_ratio
#  $ name
#  $ tradable_chain_id
#  $ splits
#  $ url
#  $ country
#  $ bloomberg_unique
#  $ list_date
}
