% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_login.R
\name{api_login}
\alias{api_login}
\title{RobinHood API: Login}
\usage{
api_login(username, password)
}
\arguments{
\item{username}{(string) RobinHood username}

\item{password}{(string) RobinHood password}
}
\description{
Backend function called by RobinHood(). Returns a list like object of class RobinHood which stores tokens required by all other functions.
}
\examples{
# data returned by api call
#  $ api_grant_type
#  $ api_client_id
#  $ tokens.access_token
#  $ tokens.refresh_token
#  $ url.positions
#  $ url.account_id
#  - attr(*, "class")= chr "RobinHood"
}
