% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_logout.R
\name{api_logout}
\alias{api_logout}
\title{Logout of RobinHood}
\usage{
api_logout(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Send a logout call through the RobinHood API service, this disables your token
}
\seealso{
\code{\link{logout}}
}
