% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_marketdata.R
\name{api_marketdata}
\alias{api_marketdata}
\title{RobinHood API: Fundamentals}
\usage{
api_marketdata(RH, instrument_id)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{instrument_id}{(string) a single instrument_id}
}
\description{
Returns a dataframe of quantitative market information for a particular option instrument
}
