% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_markets.R
\name{api_markets}
\alias{api_markets}
\title{RobinHood API: Markets}
\usage{
api_markets(RH, markets_url, type = "df")
}
\arguments{
\item{RH}{object of class RobinHood}

\item{markets_url}{(string) a single market url}

\item{type}{(string) structure of data returned, 'df' or 'list'}
}
\description{
Backend function called by get_market_hours(). Returns a data frame of markets data and trading hours.
}
\examples{
# data returned by api call
# $ website
# $ city
# $ name
# $ url
# $ country
# $ todays_hours
# $ operating_mic
# $ acronym
# $ timezone
# $ mic
}
