% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_option_positions.R
\name{api_option_positions}
\alias{api_option_positions}
\title{RobinHood API: Option Contract Positions}
\usage{
api_option_positions(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_contract_positions(). Returns a data frame of owned options contracts.
}
