% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_order_cancel.R
\name{api_order_cancel}
\alias{api_order_cancel}
\title{RobinHood API Cancel Order}
\usage{
api_order_cancel(RH, order_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{order_url}{url of order returned from place_order}
}
\description{
Send a cancel signal to robinhood
}
