% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_order_status.R
\name{api_order_status}
\alias{api_order_status}
\title{RobinHood API Order Status}
\usage{
api_order_status(RH, order_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{order_url}{url of order returned from place_order}
}
\description{
Returns the status of an order
}
