% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_orders.R
\name{api_orders}
\alias{api_orders}
\title{RobinHood API Orders}
\usage{
api_orders(RH, instrument_id, symbol, type, time_in_force, trigger, price,
  stop_price, quantity, side)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{instrument_id}{(string) URL of the instrument_id}

\item{symbol}{(string) Ticket symbol you are attempting to buy or sell}

\item{type}{(string) "market" or "limit"}

\item{time_in_force}{(string) Good For Day ("gfd"), Good Till Canceled ("gtc"), Immediate or Cancel ("ioc"), or Opening ("opg")}

\item{trigger}{(string) "immediate" or "stop"}

\item{price}{(number) the price you are willing to sell or buy at}

\item{stop_price}{(number) if trigger = stop, enter stop price, otherwise leave blank}

\item{quantity}{(int) number of shares you wish to transact}

\item{side}{(string) "buy" or "sell"}
}
\description{
Issues a buy or sell order and returns a list of order information
}
