% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_orders_options.R
\name{api_orders_options}
\alias{api_orders_options}
\title{RobinHood API: Option Orders}
\usage{
api_orders_options(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
RobinHood API: Option Orders
}
