% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_portfolio.R
\name{api_portfolios}
\alias{api_portfolios}
\title{RobinHood API: Portfolio}
\usage{
api_portfolios(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_portfolio(). Returns a data frame of account summaries.
}
