% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_portfolio_crypto.R
\name{api_portfolios_crypto}
\alias{api_portfolios_crypto}
\title{RobinHood API: Portfolio Crypto}
\usage{
api_portfolios_crypto(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_portfolio(..., source = "crypto"). Returns a data frame of the current crypto portolio summary.
}
