% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_positions.R
\name{api_positions}
\alias{api_positions}
\title{RobinHood API: Positions}
\usage{
api_positions(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_positions(). Returns a data frame of instrument position data.
}
\examples{
# data returned by api call
# $ shares_held_for_stock_grants
# $ account
# $ pending_average_buy_price
# $ shares_held_for_options_events
# $ intraday_average_buy_price
# $ url
# $ shares_held_for_options_collateral
# $ created_at
# $ updated_at
# $ shares_held_for_buys
# $ average_buy_price
# $ instrument
# $ intraday_quantity
# $ shares_held_for_sells
# $ shares_pending_from_options_events
# $ quantity
}
