% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_positions_crypto.R
\name{api_positions_crypto}
\alias{api_positions_crypto}
\title{RobinHood API: Cryptocurrency Positions}
\usage{
api_positions_crypto(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_positions(). Returns a data frame of crypto position data via the Nummus api.
}
