% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_positions_options.R
\name{api_positions_options}
\alias{api_positions_options}
\title{RobinHood API: Option Contract Positions}
\usage{
api_positions_options(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_positions_options(). Returns a data frame of owned options contracts.
}
