% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_quote.R
\name{api_quote}
\alias{api_quote}
\title{RobinHood API: Quote}
\usage{
api_quote(RH, symbols_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{symbols_url}{(string) url of query with ticker symbols}
}
\description{
Backend function called by get_positions(), get_quote(), place_order(). Returns a data frame of quote data
}
\examples{
# data returned by api call
#  $ symbol
#  $ last_trade_price
#  $ last_trade_price_source
#  $ ask_price
#  $ ask_size
#  $ bid_price
#  $ bid_size
#  $ previous_close
#  $ adjusted_previous_close
#  $ previous_close_date
#  $ last_extended_hours_trade_price
#  $ trading_halted
#  $ has_traded
#  $ updated_at
}
