% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_quote_crypto.R
\name{api_quote_crypto}
\alias{api_quote_crypto}
\title{RobinHood API: Quotes for cryptocurrency}
\usage{
api_quote_crypto(RH, symbols_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{symbols_url}{(string) url of query with ticker symbols}
}
\description{
Returns a dataframe of current quotes
}
