% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ratings.R
\name{api_ratings}
\alias{api_ratings}
\title{RobinHood API: Analyst ratings}
\usage{
api_ratings(RH, symbol)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{symbol}{(string)}
}
\description{
Backend function called by get_ratings() to get analyst rating from RobinHood
}
