% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_user.R
\name{api_user}
\alias{api_user}
\title{RobinHood API: User}
\usage{
api_user(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_user() to return user data
}
\examples{
# List of 14
# $ username
# $ first_name
# $ last_name
# $ id_info
# $ url
# $ email_verified
# $ created_at
# $ basic_info
# $ email
# $ investment_profile
# $ id
# $ international_info
# $ employment
# $ additional_info
}
