% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_watchlist.R
\name{api_watchlist}
\alias{api_watchlist}
\title{RobinHood API: watchlist}
\usage{
api_watchlist(RH, watchlist_url, detail = FALSE, delete = FALSE)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{watchlist_url}{(string) a single watchlist url}

\item{detail}{(logical) if null use header api only, otherwise pass options}

\item{delete}{(logical) send delete call}
}
\description{
Backend function called by watchlist(). Adds or remove instruments from the default watchlist. The create
and delete watchlist features are disabled as it appears that the functionality is not currently available
on the plateform.
}
