% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_ach_tranfer.R
\name{cancel_ach_transfer}
\alias{cancel_ach_transfer}
\title{Cancel an ACH transfer from your RobinHood account}
\usage{
cancel_ach_transfer(RH, cancel_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{cancel_url}{(string) cancel url returned from initiating a transfer with place_ach_transfer()}
}
\description{
Cancel an ACH transfer from your RobinHood account
}
