% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rh.R
\name{check_rh}
\alias{check_rh}
\title{Checks to see if you have a valid RobihHood object}
\usage{
check_rh(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Checks to see if you have a valid RobihHood object
}
