% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accounts.R
\name{get_accounts}
\alias{get_accounts}
\title{Get account data}
\usage{
get_accounts(RH)
}
\arguments{
\item{RH}{object class RobinHood}
}
\description{
Download a list of account data for your RobinHood account.
}
