% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accounts_list.R
\name{get_accounts_list}
\alias{get_accounts_list}
\title{Get data related to your RobinHood account in list format}
\usage{
get_accounts_list(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Get data related to your RobinHood account in list format
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password", mfa_code = '123456')

get_accounts_list(RH)
}
}
