% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ach.R
\name{get_ach}
\alias{get_ach}
\title{Get ACH data from your RobinHood linked bank accounts}
\usage{
get_ach(RH, action, status_url = NULL)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "transfers", "relationships", "schedules", "status"}

\item{status_url}{(string) URL returned by place_ach_transfer()}
}
\description{
Get ACH data from your RobinHood linked bank accounts
}
