% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contracts.R
\name{get_contracts}
\alias{get_contracts}
\title{Get a option contracts from RobinHood}
\usage{
get_contracts(RH, chain_symbol, type, detail = FALSE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{chain_symbol}{(string) a single ticket symbol}

\item{type}{(string) one of call or put}

\item{detail}{(logical) if TRUE (default) return additional info on greeks, prevous day, high/low fill rate prices}
}
\description{
Get a option contracts from RobinHood
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_contracts(RH, "IR")
}
}
