% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_hours.R
\name{get_market_hours}
\alias{get_market_hours}
\title{Get market hours for a specific date}
\usage{
get_market_hours(RH, market_date = NULL)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{market_date}{(string) date in the form 'yyyy-mm-dd', default today}
}
\description{
Get a list of markets
}
