% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_hours.R
\name{get_nyse}
\alias{get_nyse}
\title{Get NYSE Hours}
\usage{
get_nyse(RH, market_date = Sys.Date())
}
\arguments{
\item{RH}{object of class RobinHood}

\item{market_date}{(string) date in the form 'yyyy-mm-dd', default today}
}
\description{
Get NYSE Hours in New York Hours - Forced to be NYSE to accommodate Robinhood error
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_nyse(RH)
}
}
