% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_history.R
\name{get_order_history}
\alias{get_order_history}
\title{Download all available order history for your RobinHood account}
\usage{
get_order_history(RH, page_size = 1000)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{page_size}{(int) number of historical records to fetch}
}
\description{
Download all available order history for your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_order_history(RH)
}
}
