% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_history_options.R
\name{get_order_history_options}
\alias{get_order_history_options}
\title{Download all available options order history for your RobinHood account}
\usage{
get_order_history_options(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Download all available options order history for your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_order_history(RH)
}
}
