% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolios.R
\name{get_portfolios}
\alias{get_portfolios}
\title{Get porfolio summaries related to your RobinHood account}
\usage{
get_portfolios(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Get porfolio summaries related to your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_portfolios(RH)
}
}
