% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions.R
\name{get_positions}
\alias{get_positions}
\title{Get the currently held positions for your RobinHood account}
\usage{
get_positions(RH, limit_output = TRUE, options = FALSE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{limit_output}{(logical) if true, return a simplified positions table, false returns all position details}

\item{options}{(logical) if true, include any options or pending options}
}
\description{
Get the currently held positions for your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_positions(RH)
}
}
