% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions_crypto.R
\name{get_positions_crypto}
\alias{get_positions_crypto}
\title{Get the currently held crypto positions for your RobinHood account}
\usage{
get_positions_crypto(RH)
}
\arguments{
\item{RH}{object class RobinHood}
}
\description{
Get the currently held crypto positions for your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_positions_crypto(RH)
}
}
