% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions_options.R
\name{get_positions_options}
\alias{get_positions_options}
\title{Get the currently held positions for your RobinHood account}
\usage{
get_positions_options(RH, trim_pending = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{trim_pending}{(logical) if FALSE, then return pending and intraday columns}
}
\description{
Get the currently held positions for your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_positions_options(RH)
}
}
