% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote_crypto.R
\name{get_quote_crypto}
\alias{get_quote_crypto}
\title{Get a crypto currency quote from RobinHood}
\usage{
get_quote_crypto(RH, symbol)
}
\arguments{
\item{RH}{object class RobinHood}

\item{symbol}{(string) cryto currency symbol such as BTC, ETH, DOGE}
}
\description{
Get a crypto currency quote from RobinHood
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_quote_crypto(RH, "BTC")
}
}
