% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratings.R
\name{get_ratings}
\alias{get_ratings}
\title{Get analyst raings for a given equity symbol from RobinHood}
\usage{
get_ratings(RH, symbol)
}
\arguments{
\item{RH}{object class RobinHood}

\item{symbol}{(string) Ticker symbol}
}
\description{
Get analyst raings for a given equity symbol from RobinHood
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_ratings(RH)
}
}
