% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag.R
\name{get_tag}
\alias{get_tag}
\title{Get ticker symbols for popular tags on RobinHood}
\usage{
get_tag(RH, tag)
}
\arguments{
\item{RH}{object class RobinHood}

\item{tag}{(string) a hyphenated tag such as "100-most-popular"}
}
\description{
Geta a data frame of ticker symbols and names linked to trending tags on the RobinHood website.
}
