% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Get tickers}
\usage{
get_tickers(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Get a dataframe of all listed companies including tickers and trade indicators.
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_tickers(RH)
}
}
