% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Get user data}
\usage{
get_user(RH, ticker, limit_output = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}
}
\description{
Download a list of user data from your RobinHood account
}
