% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositions.R
\name{getPositions}
\alias{getPositions}
\title{Get Current Positions of your portfolio}
\usage{
getPositions(RH)
}
\arguments{
\item{RH}{object class RobinHood}
}
\description{
Get Current Positions of your portfolio
}
\examples{
# Get you current positions
# RH <- RobinHood(username = 'your username', password = 'your password')
# getPositions(RH)
}
