% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURL.R
\name{getURL}
\alias{getURL}
\title{Get the URL for the RobinHood API Endpoints}
\usage{
getURL(endpoint)
}
\arguments{
\item{endpoint}{which api endpoint to look up?}
}
\description{
Get the URL for the RobinHood API Endpoints
}
