% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_json.R
\name{mod_json}
\alias{mod_json}
\title{Converts a data frame to json minus the square brackets on the ends}
\usage{
mod_json(x, type)
}
\arguments{
\item{x}{dataframe intended for json conversoin}

\item{type}{(string) one of "fromJSON" or "toJSON"}
}
\description{
Converts a data frame to json minus the square brackets on the ends
}
