% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_ach_transfer.R
\name{place_ach_transfer}
\alias{place_ach_transfer}
\title{Place an ACH transfer to and from your RobinHood account}
\usage{
place_ach_transfer(RH, action, amount, transfer_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "deposit", "withdraw"}

\item{amount}{(numeric) amount in dollars you want to deposit or withdraw}

\item{transfer_url}{(string) url of your linked account, output of get_ach(RH, "relationships")}
}
\description{
Place an ACH transfer to and from your RobinHood account
}
