% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobinHood.R
\name{RobinHood}
\alias{RobinHood}
\title{RobinHood Account Authentication}
\usage{
RobinHood(
  username,
  password,
  mfa_code = "NA",
  totp_code = "NA",
  account_type = "individual"
)
}
\arguments{
\item{username}{(string) account email address}

\item{password}{(string) password}

\item{mfa_code}{(string) mfa_code provided by your authentication app (required if mfa is enabled)}

\item{totp_code}{(string) the Code Robinhood provides to set in the authenticator app}

\item{account_type}{(string) for Retirement accounts, use 'ira_roth' or 'ira_traditional'}
}
\description{
This function returns an object of S3 class RobinHood and establishes a connection to a RobinHood account.
It is a required input for every other function in the package.
}
\examples{
\dontrun{

## Logging into the primary account using an mfa code from authenticator
RH <- RobinHood("username", "password", mfa_code = "123456")

## Alternatively, this allows the mfa code to be generated automatically using the code
## Robinhood provides when you set up two factor authentication. This also logs into an IRA account
RH <- RobinHood("username", "password", totp_code = "JBSWY3DPEHPK3PXP", account_type = "ira_roth")

}
}
