% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_accounts.R
\name{api_accounts}
\alias{api_accounts}
\title{RobinHood API: User}
\usage{
api_accounts(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_user(),'returns a list of user account data.
}
