% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ach.R
\name{api_ach}
\alias{api_ach}
\title{RobinHood API: ACH}
\usage{
api_ach(RH, action)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "history", "accounts", "schedules"}
}
\description{
Backend function for interacting and getting data on linked bank accounts.
}
