% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ach.R
\name{api_ach}
\alias{api_ach}
\title{RobinHood API: ACH}
\usage{
api_ach(RH, action, status_url = NULL, cancel_url = NULL)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "transfers", "relationships", "schedules", "status", "cancel"}

\item{status_url}{(string) URL returned by place_ach_transfer()}

\item{cancel_url}{(string) URL returned by place_ach_transfer()}
}
\description{
Backend function for interacting and getting data on linked bank accounts.
}
