% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints.R
\name{api_endpoints}
\alias{api_endpoints}
\title{RobinHood API: Endpoints}
\usage{
api_endpoints(endpoint)
}
\arguments{
\item{endpoint}{(string) which api endpoint to look up?}
}
\description{
Returns the appropriate starting URL for a given endpoint.
}
