% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fundamentals.R
\name{api_fundamentals}
\alias{api_fundamentals}
\title{RobinHood API: Fundamentals}
\usage{
api_fundamentals(RH, ticker)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{ticker}{(string) vector of ticker symbols}
}
\description{
Get a list of trading data for a particular ticker symbol
}
