% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fundamentals.R
\name{api_fundamentals}
\alias{api_fundamentals}
\title{RobinHood API: Fundamentals}
\usage{
api_fundamentals(RH, ticker)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{ticker}{(string) vector of ticker symbols}
}
\description{
Backend function called by get_fundamentals(), watchlist(). Returns a data frame of descriptive data for a given ticker symbol.
}
\examples{
# data returned by api call
# $ open
# $ high
# $ low
# $ volume
# $ average_volume_2_weeks
# $ average_volume
# $ high_52_weeks
# $ dividend_yield
# $ low_52_weeks
# $ market_cap
# $ pe_ratio
# $ shares_outstanding
# $ description
# $ instrument
# $ ceo
# $ headquarters_city
# $ headquarters_state
# $ sector
# $ industry
# $ num_employees
# $ year_founded
}
