% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_instruments.R
\name{api_instruments}
\alias{api_instruments}
\title{RobinHood API: Instruments}
\usage{
api_instruments(RH, instrument_id)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{instrument_id}{(string) a single instrument_id}
}
\description{
Backend function called by get_tag(), get_position(), watchlist(). Returns a list of instrument data.
}
