% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_instruments.R
\name{api_instruments}
\alias{api_instruments}
\title{RobinHood API Instruments}
\usage{
api_instruments(instrument_id)
}
\arguments{
\item{instrument_id}{(string) instrument_id returned from api_positions}
}
\description{
Returns a list of instrument data
}
