% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_instruments.R
\name{api_instruments}
\alias{api_instruments}
\title{RobinHood API: Instruments}
\usage{
api_instruments(RH, symbol = NULL, instrument_url = NULL)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{symbol}{(string) a single symbol}

\item{instrument_url}{(string) instrument url}
}
\description{
Backend function called by get_tag(), get_position(), watchlist(). Returns a list of instrument data.
}
