% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_login.R
\name{api_login}
\alias{api_login}
\title{RobinHood API: Login}
\usage{
api_login(username, password)
}
\arguments{
\item{username}{RobinHood username}

\item{password}{RobinHood password}
}
\description{
Returns a list like object of class RobinHood
}
