% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_logout.R
\name{api_logout}
\alias{api_logout}
\title{RobinHood API: Logout}
\usage{
api_logout(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by logout(). Sends a logout call and disables your oauth2 token.
}
