% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_marketdata.R
\name{api_marketdata}
\alias{api_marketdata}
\title{RobinHood API: Market Data}
\usage{
api_marketdata(RH, instrument, type = "instrument_id")
}
\arguments{
\item{RH}{object of class RobinHood}

\item{instrument}{(string) a single instrument_id or multiple instrument_urls}

\item{type}{(string) one of instrument_id or instrument_url}
}
\description{
Returns a dataframe of quantitative market information for a particular option instrument
}
