% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_markets.R
\name{api_markets}
\alias{api_markets}
\title{RobinHood API: Markets}
\usage{
api_markets(RH, markets_url, type = "df")
}
\arguments{
\item{RH}{object of class RobinHood}

\item{markets_url}{(string) url of the market api}

\item{type}{(string) 'df' or 'list'}
}
\description{
Get a list of markets
}
