% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_positions.R
\name{api_positions}
\alias{api_positions}
\title{RobinHood API: Positions}
\usage{
api_positions(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_positions(). Returns a data frame of instrument position data.
}
