% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_quote.R
\name{api_quote}
\alias{api_quote}
\title{RobinHood API: Quote}
\usage{
api_quote(RH, symbols_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{symbols_url}{(string) url of query with ticker symbols}
}
\description{
Returns a list of instrument data
}
