% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tag.R
\name{api_tag}
\alias{api_tag}
\title{RobinHood API: Tag}
\usage{
api_tag(RH, tag)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{tag}{(string) a hyphenated tag such as "100-most-popular"}
}
\description{
Backend function called by get_tag(). Returns a list of instrument ids.
}
