% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tickers.R
\name{api_tickers}
\alias{api_tickers}
\title{RobinHood API: Tickers}
\usage{
api_tickers(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Backend function called by get_ticker. Returns a data frame of all instruments listed
on RobinHood.
}
