% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_order.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel an existing order on RobinHood}
\usage{
cancel_order(RH, order_url)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{order_url}{(string) url of order returned from place_order()}
}
\description{
Send a cancel buy order command to RobinHood. You will need to retain the buy/sell order url.
}
