% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accounts.R
\name{get_accounts}
\alias{get_accounts}
\title{Get data related to your RobinHood account}
\usage{
get_accounts(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Get data related to your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_accounts(RH)
}
}
