% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ach.R
\name{get_ach}
\alias{get_ach}
\title{Get ACH data from your RobinHood linked bank accounts}
\usage{
get_ach(RH, action)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{action}{(string) one of "transfers", "relationships", "schedules"}
}
\description{
Get ACH data from your RobinHood linked bank accounts
}
