% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fundamentals.R
\name{get_fundamentals}
\alias{get_fundamentals}
\title{Get fundamentals for a particular symbol or list of symbols}
\usage{
get_fundamentals(RH, ticker, include_description = FALSE)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{ticker}{(string) vector of ticker symbols}

\item{include_description}{(logical) include a long description of the company (default: FALSE)}
}
\description{
Get fundamentals
}
