% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fundamentals.R
\name{get_fundamentals}
\alias{get_fundamentals}
\title{Get ticker fundamentals}
\usage{
get_fundamentals(RH, ticker, include_description = FALSE)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{ticker}{(string) vector of ticker symbols}

\item{include_description}{(logical) include a long description of the company (default: FALSE)}
}
\description{
Get ticker fundamentals for a particular symbol or list of symbols.
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_fundamentals(RH, "CAT")
}
}
