% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_history.R
\name{get_order_history}
\alias{get_order_history}
\title{Get order history using the RobinHood api}
\usage{
get_order_history(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Returns recent order history.
}
