% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_status.R
\name{get_order_status}
\alias{get_order_status}
\title{Get order status using the RobinHood api}
\usage{
get_order_status(RH, order_url, limit_output = TRUE)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{order_url}{(string) url of order returned from place_order}

\item{limit_output}{(logical) return limited info on the order (default TRUE)}
}
\description{
Returns a list of order information given a buy/sell order url.
}
