% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions.R
\name{get_positions}
\alias{get_positions}
\title{Get the current positions of your account}
\usage{
get_positions(RH, limit_output = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{limit_output}{(logical) if true, return a simplified positions table, false returns all positions details}
}
\description{
Download all investment positions for your account.
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_positions(RH)
}
}
