% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions.R
\name{get_positions}
\alias{get_positions}
\title{Get Current Positions of your portfolio}
\usage{
get_positions(RH, simple = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{simple}{(logical) if true, return a simplified positions table, false returns all positions details}
}
\description{
Get Current Positions of your portfolio
}
\examples{
# Get you current positions
# RH <- RobinHood(username = 'your username', password = 'your password')
# get_positions(RH)
}
