% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote.R
\name{get_quote}
\alias{get_quote}
\title{Get a quote from RobinHood}
\usage{
get_quote(RH, ticker, limit_output = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{ticker}{(string) of ticker symbols}

\item{limit_output}{(logical) if TRUE (default) return less quote detail}
}
\description{
Get a quote from RobinHood
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_quote(RH, "IR")
}
}
