% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote.R
\name{get_quote}
\alias{get_quote}
\title{Get a current instrument quote from RobinHood}
\usage{
get_quote(RH, ticker, simple = TRUE)
}
\arguments{
\item{RH}{object class RobinHood}

\item{ticker}{(string) of ticker symbols}
}
\description{
For a string of ticker symbols, return quote data.
}
\examples{
# Get you current positions
# RH <- RobinHood(username = 'your username', password = 'your password')
# get_quote(RH, c("CAT", "GE"))
}
