% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickers.R
\name{get_tickers}
\alias{get_tickers}
\title{Download all financial instruments available on RobinHood}
\usage{
get_tickers(RH, fundamentals = FALSE)
}
\arguments{
\item{RH}{object of class RobinHood}

\item{fundamentals}{(logical) if TRUE then return fundamental data (long run time)}
}
\description{
Get a dataframe of all listed companies including ticker symbols and tradeability indicators.
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_tickers(RH)
}
}
