% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Get personal user data related to your RobinHood account}
\usage{
get_user(RH)
}
\arguments{
\item{RH}{object class RobinHood}
}
\description{
Get personal user data related to your RobinHood account
}
\examples{
\dontrun{
# Login in to your RobinHood account
RH <- RobinHood("username", "password")

get_user(RH)
}
}
