% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositions.R
\name{getPositions}
\alias{getPositions}
\title{Get Current Positions of your portfolio}
\usage{
getPositions(rh)
}
\arguments{
\item{rh}{object class RobinHood}
}
\description{
Get Current Positions of your portfolio
}
\examples{
# Get you current positions
getPositions(rh)
}
