% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout of RobinHood}
\usage{
logout(RH)
}
\arguments{
\item{RH}{object of class RobinHood}
}
\description{
Send a logout call through the RobinHood API service and disable your token.
}
\examples{
# Logout
# RH <- RobinHood(username = 'your username', password = 'your password')
# logout(RH)
}
