% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobinHood.R
\name{RobinHood}
\alias{RobinHood}
\title{Connect to your RobinHood Account}
\usage{
RobinHood(username, password)
}
\arguments{
\item{username}{user name or email}

\item{password}{password}
}
\description{
This function returns an object of S3 class RobinHood and establishes a
connection to a RobinHood account.
}
\examples{
# Connect to your robinhood account
# RH <- RobinHood(username = 'your username', password = 'your password')
}
