% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchlist.R
\name{watchlist}
\alias{watchlist}
\title{Get a current instrument quote from RobinHood}
\usage{
watchlist(RH, action, watchlist = NULL, ticker = NULL)
}
\arguments{
\item{RH}{object class RobinHood}

\item{action}{(string) one of: get, add, delete}

\item{watchlist}{(string) name of watchlist to add, delete, or get instruments, null will return a list of watchlist}

\item{ticker}{(string) list of tickers to add or delete, null will add or delete watchlist}
}
\description{
For a string of ticker symbols, return quote data.
}
\examples{
# Get you current positions
# RH <- RobinHood(username = 'your username', password = 'your password')
# get_quote(RH, c("CAT", "GE"))
}
